Initial draft of Contributing docs. (#170)
This is an initial draft of the Contributing docs. It includes an overview of how to contribute, how to document a test run, a statement about testing on your own account, and how to set up a development environment. It also reorganizes the quick start guides into one directory, instead of being spread out in platform-specific directories. There's also an issue template for recording test run reports.